+++ 08_audioExamples/01_usingJack/05_synthInheritence/simpleSynth.h
@@ -7,8 +7,8 @@
 class SimpleSynth : public Synth {
 public:
   //constructor
-  SimpleSynth(double samplerate);
-  SimpleSynth(double samplerate, float midiPitch);
+  SimpleSynth(float samplerate);
+  SimpleSynth(float samplerate, float midiPitch);
   //prevent the default constructor to be generated
   SimpleSynth() = delete;
   //destructor
